# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Job Application Tracker API",
        "version": "1.0.0",
        "description": "REST API for managing companies, job applications, contacts, and interviews."
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
...
    "paths": {
...
        "/interviews": {
            "get": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Get all interviews",
                "operationId": "getAllInterviews",
                "responses": {
                    "200": {
                        "description": "Interviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Interview"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Create a interview",
                "operationId": "createInterview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InterviewInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Interview created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Interview"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
        "/interviews/{id}": {
            "parameters": [
                {
                    "$ref": "#/components/parameters/DocumentId"
                }
            ],
            "get": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Get one interview",
                "operationId": "getInterviewById",
                "responses": {
                    "200": {
                        "description": "Interview retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Interview"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            },
            "put": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Update a interview",
                "operationId": "updateInterview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InterviewInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interview updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Interview"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Delete a interview",
                "operationId": "deleteInterview",
                "responses": {
                    "204": {
                        "description": "Interview deleted successfully"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
        "/auth/github": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign in with GitHub",
                "description": "Open /auth/github in a browser tab to complete login. Swagger Execute cannot complete the interactive redirect flow.",
                "operationId": "loginWithGitHub",
                "responses": {
                    "302": {
                        "description": "Redirect to GitHub for authorization"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
        "/auth/status": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Check current login session",
                "operationId": "getAuthenticationStatus",
                "responses": {
                    "200": {
                        "description": "Session status; user is returned when authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthStatus"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out and destroy the session",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "DocumentId": {
                "name": "id",
                "in": "path",
                "required": true,
                "description": "MongoDB document ID",
                "schema": {
                    "type": "string",
                    "pattern": "^[a-fA-F0-9]{24}$"
                }
            }
        },
        "responses": {
            "BadRequest": {
                "description": "Invalid ID or request data",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Document not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "An authenticated GitHub session is required",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/AuthenticationError"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unable to process the request"
                    }
                }
            },
...
            "InterviewInput": {
                "type": "object",
                "required": [
                    "applicationId",
                    "scheduledAt",
                    "type",
                    "interviewer",
                    "location",
                    "status"
                ],
                "properties": {
                    "applicationId": {
                        "type": "string",
                        "example": "6aab443d3f32b48f16367f8e",
                        "minLength": 1,
                        "pattern": "^[a-fA-F0-9]{24}$",
                        "description": "ID of an existing job application"
                    },
                    "scheduledAt": {
                        "type": "string",
                        "example": "2026-10-20T15:00:00Z",
                        "minLength": 1,
                        "format": "date-time",
                        "description": "ISO date-time including seconds and timezone (Z or an offset).",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$"
                    },
                    "type": {
                        "type": "string",
                        "example": "video",
                        "minLength": 1,
                        "enum": [
                            "phone",
                            "video",
                            "onsite"
                        ]
                    },
                    "interviewer": {
                        "type": "string",
                        "example": "Jordan Smith",
                        "minLength": 1
                    },
                    "location": {
                        "type": "string",
                        "example": "https://example.com/meeting",
                        "minLength": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "scheduled",
                        "minLength": 1,
                        "enum": [
                            "scheduled",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Test interview for Junior Web Developer"
                    }
                }
            },
            "Interview": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/InterviewInput"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string",
                                "example": "68cb1234567890abcdef1234"
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                ]
            },
            "AuthenticationError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Authentication required"
                    },
                    "login": {
                        "type": "string",
                        "example": "/auth/github"
                    }
                }
            },
            "AuthStatus": {
                "type": "object",
                "properties": {
                    "authenticated": {
                        "type": "boolean"
                    },
                    "login": {
                        "type": "string",
                        "example": "/auth/github"
                    },
                    "user": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "username": {
                                "type": "string"
                            },
                            "displayName": {
                                "type": "string"
                            },
                            "profileUrl": {
                                "type": "string"
                            },
                            "avatarUrl": {
                                "type": "string",
                                "nullable": true
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "job-tracker.sid",
                "description": "Sign in through /auth/github in this browser. The session cookie is sent automatically to same-origin requests."
            }
        }
    }
}